ADL: Clean up AdlMetaEngine
